logika-server/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
sqlalchemy>=2,<3
psycopg2-binary>=2.9,<3
pydantic>=2,<3
pydantic-settings>=2,<3
httpx>=0.27,<1.0
anthropic>=0.45,<1.0
pyjwt[crypto]>=2.9,<3
reportlab>=4.2,<5
